lab/이인재/규격/requirements.txt
asyncpg==0.30.0
sentence-transformers==3.3.1
pymupdf4llm==0.0.17
langchain-text-splitters==0.3.2
openai==1.57.2
python-dotenv==1.0.1
pandas==2.2.3
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.24.0
